Tests for the full screen edit mode
These aren't ideal as it's based on html extracted from gcal. But better than nothing, and provides a framework for diagnosing issues going forward. Due to the very dynamic nature of gcal, mnost likely issues will need to be reproduced "live" but then can be incorporated into this framework to avoid future regression.