Ensure "Add Brave Talk" button gets added to Quick Add
It looks like the logic of this dialog has changed, in that after initail display it now gets rewritten. So the button _was_ being added, but then subequently removed.

The "fix" here is to delay a bit before trying to add the button, which generates a bit of jumpiness but seems to work reliably for me.

Fixes brave/brave-talk-gcalendar-extension#21